Chaning Tabulated to have ints for boundaries and schemes.

+++ src/ACEtk/interpretation/ContinuousEnergyNeutron/Builder/FissionMultiplicity/Tabulated/Builder/test/Builder.test.cpp
@@ -20,8 +20,8 @@ SCENARIO( "Testing FissionMultiplicity::Tabulated::Builder" ){
       using TabulatedBuilder::TabulatedBuilder;
     };
 
-    std::vector< double > boundaries{ 0.0, 3.0 };
-    std::vector< double > schemes{ 2.0, 1.0 };
+    std::vector< int > boundaries{ 0, 3 };
+    std::vector< int > schemes{ 2, 1 };
     std::vector< double > energies{ 1.0, 2.0, 5.0, 6.0 };
     std::vector< double > multiplicities{ 2.1, 2.2, 2.5, 2.5 };
     TestBuilder tb{ grandparentBuilder, nuType };
